cloud_storage: add bazel tests
